+++ src/integrationTest/java/uk/gov/hmcts/reform/em/ped/config/SwaggerPublisherTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.demo.config;
+package uk.gov.hmcts.reform.em.ped.config;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
