+++ src/main/java/uk/gov/hmcts/reform/em/ped/config/StompUnsubscribedEvent.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.demo.config;
+package uk.gov.hmcts.reform.em.ped.config;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -6,26 +6,26 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationListener;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.stereotype.Component;
-import org.springframework.web.socket.messaging.SessionDisconnectEvent;
-import uk.gov.hmcts.reform.demo.services.ParticipantsStatusService;
+import org.springframework.web.socket.messaging.SessionUnsubscribeEvent;
+import uk.gov.hmcts.reform.em.ped.services.ParticipantsStatusService;
 
 @Component
-public class StompDisconnectedEvent implements ApplicationListener<SessionDisconnectEvent> {
+public class StompUnsubscribedEvent implements ApplicationListener<SessionUnsubscribeEvent> {
 
-    private static final Logger LOG = LoggerFactory.getLogger(StompDisconnectedEvent.class);
+    private static final Logger LOG = LoggerFactory.getLogger(StompUnsubscribedEvent.class);
 
     private final ParticipantsStatusService participantsStatusService;
 
     @Autowired
-    public StompDisconnectedEvent(ParticipantsStatusService participantsStatusService) {
+    public StompUnsubscribedEvent(ParticipantsStatusService participantsStatusService) {
         this.participantsStatusService = participantsStatusService;
     }
 
     @Override
-    public void onApplicationEvent(SessionDisconnectEvent event) {
+    public void onApplicationEvent(SessionUnsubscribeEvent event) {
         StompHeaderAccessor sha = StompHeaderAccessor.wrap(event.getMessage());
         String sessionId = sha.getSessionId();
         LOG.info(sessionId);
-        participantsStatusService.removeParticipant(sessionId);
+        participantsStatusService.unsubscribeParticipant(sessionId);
     }
 }
