+++ src/main/java/uk/gov/hmcts/reform/em/ped/controllers/UpdateController.java
@@ -1,9 +1,9 @@
-package uk.gov.hmcts.reform.demo.controllers;
+package uk.gov.hmcts.reform.em.ped.controllers;
 
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.stereotype.Controller;
-import uk.gov.hmcts.reform.demo.domain.ScreenChange;
+import uk.gov.hmcts.reform.em.ped.domain.ScreenChange;
 
 @Controller
 public class UpdateController {
